+++ CHANGES.rst
@@ -11,6 +11,10 @@ Version 1.10
 
 Unreleased
 
+* Use indexes for :meth:`~Reader.get_entries()` (recent order);
+  should make calls 10-30% faster.
+  (:issue:`134`)
+
 
 Version 1.9
 -----------